requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
importlib-metadata==3.1.0
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
